backend-ai/requirements.txt
cryptography==46.0.7
fastapi==0.136.0
google-generativeai==0.8.6
httpx==0.28.1
python-dotenv==1.2.2
python-jose==3.5.0
supabase==2.28.3
uvicorn==0.44.0
